move make_invalid_op to ops
